Removing Table container component, Table presenter now gets result props directly from App component
